+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## v1.4.2
+- Fixed French translations.
+- Fixed bug where days of the week of weekly recurrence were not checked.
+- Target API changed to 28.
+
 ## v1.4.1
 - Fixed library dependencies not being added to the project